Move unit tests into tests/ folder
Relocate the colocated src unit tests to a dedicated tests/ directory
(matching the e2e/ convention), update the vp test script to run tests/,
and add tests/ to the tsconfig include.
