+++ packages/react-grab/tsconfig.json
@@ -10,6 +10,13 @@
     "lib": ["esnext", "dom", "dom.iterable"],
     "skipLibCheck": true
   },
-  "include": ["src", "vite.config.ts", "solid-babel-plugin.ts", "e2e", "playwright.config.ts"],
+  "include": [
+    "src",
+    "tests",
+    "vite.config.ts",
+    "solid-babel-plugin.ts",
+    "e2e",
+    "playwright.config.ts"
+  ],
   "exclude": ["**/node_modules/**", "dist"]
 }
